Move reproduction test to standalone file
Avoids the fetch_url import conflict between upstream/main test_upload.py
and the installed utils.py in the dev container.
